+++ src/main/java/io/github/mortuusars/monobank/integration/thief/ThiefCrime.java
@@ -1,0 +1,8 @@
+package io.github.mortuusars.monobank.integration.thief;
+
+public enum ThiefCrime {
+    NONE,
+    LIGHT,
+    MEDIUM,
+    HEAVY;
+}