+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 1.1.5 - 2025-05-20
+- Removed all thief-related functionality in favor of integration with `Thief` mod.
+- Monobanks are no longer carryable with `Carry On` mod.
+- Fixed 'Break In Attempted' or 'Break In Successful' warnings being resettable by a player that does not own the Monobank.
+- Break In warnings no longer show for a player that does not own the Monobank.
+
 ### 1.1.4 - 2024-02-07
 - Fixed Monobanks in most villages not having a loot [1.20.1]